sc/requirements.txt
asgiref==3.2.3
Django==3.0
django-filter==2.0.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Markdown==3.1.1
Pillow==7.0.0
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
uuid==1.30
xlrd==1.2.0